all tests passed just need to perfect my code

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: scosta-j <scosta-j@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/26 19:05:51 by scosta-j          #+#    #+#             */
-/*   Updated: 2023/01/26 22:32:05 by scosta-j         ###   ########.fr       */
+/*   Updated: 2023/01/27 18:26:32 by scosta-j         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int		print_unsigned(va_list args);
 int		compute_size_u(unsigned int n);
 int		print_hexadecimal(va_list args, char type);
 int		compute_size_hex(unsigned long long int n);
-int		turning_hexa(unsigned long long int a, char *base);
+int		turning_hexa(unsigned long int a, char *base);
 int		print_pointer(va_list args);
 
 #endif
